% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abx_days}
\alias{abx_days}
\title{Antibiotic usage.}
\format{
A data frame with 8441 rows and 5 variables:
\describe{
\item{hosp}{Hospital, character}
\item{month}{Month, date}
\item{atc}{ATC code, character}
\item{drug}{Drug name, character}
\item{days}{Number of antibiotic treatment days, 1 patient x 1 drug = 1 day}
}
}
\usage{
abx_days
}
\description{
A dataset containing antibiotic usage data from from six hospitals from
varying time periods.
}
\keyword{datasets}
