% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abx_sales}
\alias{abx_sales}
\title{AWaRe classification of ATC codes.}
\format{
A data frame with 47551 rows and 7 variables:
\describe{
\item{region}{Region, character}
\item{hospital}{Hospital, character}
\item{month}{Month, date}
\item{atc}{ATC code, character}
\item{drug}{Drug name, character}
\item{ddd}{Amount purchased, Defined Daily Dosages (DDD)}
\item{addd}{Amount purchased, DDDs adjusted to Danish recommendations}
...
}
}
\source{
\url{https://amgros.dk/}
}
\usage{
abx_sales
}
\description{
A dataset containing antibiotic sales data from Danish hospitals 2015-2020.
}
\keyword{datasets}
