% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abxaware-package.R
\docType{package}
\name{abxaware-package}
\alias{abxaware}
\alias{abxaware-package}
\title{abxaware: AWaRe classification of Antibiotic Usage Data}
\description{
Aggregate and plot antibiotic usage data by access, watch, and reserve (AWaRe) groups.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://anhoej.github.io/abxaware/}
}

}
\author{
\strong{Maintainer}: Jacob Anhøj \email{jacob@anhoej.net} (\href{https://orcid.org/0000-0002-7701-1774}{ORCID})

Authors:
\itemize{
  \item Jonas Bredtoft Boel (\href{https://orcid.org/0000-0003-3085-0930}{ORCID})
}

}
\keyword{internal}
