% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awr_funs.R
\name{awr_aggregate}
\alias{awr_aggregate}
\title{Aggregate antibiotic usage data by AWaRe group}
\usage{
awr_aggregate(
  df,
  atc = atc,
  ddd = ddd,
  ...,
  tall = FALSE,
  method = c("dk", "who"),
  ignore.other = FALSE
)
}
\arguments{
\item{df}{Data frame.}

\item{atc}{ATC code.}

\item{ddd}{Amount (usually Defined Daily Doses).}

\item{...}{Grouping variables.}

\item{tall}{If TRUE (default) outputs data in tall format.}

\item{method}{'dk' (default) or 'who' indicating the AWaRe classification to
be used.}

\item{ignore.other}{If TRUE, ignores drugs that have no AWaRe class.}
}
\value{
A data frame.
}
\description{
Aggregate antibiotic usage data by AWaRe group
}
\examples{
awr_aggregate(abx_sales)
awr_aggregate(abx_sales, atc, ddd)
awr_aggregate(abx_sales, atc, ddd, region)
awr_aggregate(abx_sales, atc, ddd, region, tall = TRUE)
awr_aggregate(abx_sales, atc, ddd, month)
awr_aggregate(abx_sales, atc, ddd, month, region)
awr_aggregate(abx_sales, atc, ddd, month, region, hospital)

}
