% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awr_funs.R
\name{awr_plot}
\alias{awr_plot}
\title{Plot AWaRe data}
\usage{
awr_plot(
  df,
  atc = atc,
  ddd = ddd,
  time = NULL,
  unit = NULL,
  ncol = NULL,
  legend.position = NULL,
  runs.analysis = TRUE,
  ...
)
}
\arguments{
\item{df}{Data frame.}

\item{atc}{ATC code.}

\item{ddd}{Amount.}

\item{time}{Time period.}

\item{unit}{Organisational unit.}

\item{ncol}{Integer, number of columns in faceted plots.}

\item{legend.position}{Character, where to put legend (e.g. 'none', 'right',
'bottom').}

\item{runs.analysis}{Logical, if TRUE (default) add runs analysis to plot.
See Details.}

\item{...}{Other arguments to \code{\link{awr_aggregate}} and
\code{\link[ggplot2]{ggplot}}.}
}
\value{
A ggplot object.
}
\description{
Plot AWaRe data
}
\details{
The plot show the relative distribution of antibiotics in the three
AWaRe classes as suggested by WHO (\url{https://adoptaware.org/}): green =
access, amber = watch, red = reserve.

If \code{runs.analysis = TRUE} (default) and \code{time} is provided, the
plot will include a horisontal line representing the median of the access
proportion. If the runs analysis finds non-random variation in the form of
either unusually long runs of data points on the same side of the centre
line or unusually few crossing of the centre line
(Anhøj 2014, \doi{10.1371/journal.pone.0113825}), the line will be dashed, otherwise
the line is solid.
}
\examples{
awr_plot(abx_sales, atc, ddd)
awr_plot(abx_sales, atc, ddd, unit = region)
awr_plot(abx_sales, atc, ddd, time = month)
awr_plot(abx_sales, atc, ddd, time = month, unit = region)
awr_plot(abx_sales, atc, ddd, time = month, unit = hospital, ncol = 4)

}
