% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abx_aware}
\alias{abx_aware}
\title{AWaRe classification of ATC codes.}
\format{
A data frame with 318 rows and 3 variables:
\describe{
\item{atc}{ATC code, character}
\item{text}{Drug name, character}
\item{aware}{AWaRe group, access, watch or reserve, factor}
...
}
}
\source{
\url{https://adoptaware.org/}
}
\usage{
abx_aware
}
\description{
A dataset containing ATC codes and the corresponding AWaRe categories.
}
\keyword{datasets}
