% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abx_days}
\alias{abx_days}
\title{Antibiotic usage data from six hospitals.}
\format{
A data frame with 8441 rows and 5 variables:
\describe{
\item{hosp}{Hospital, character}
\item{month}{Month, date}
\item{atc}{ATC code, character}
\item{drug}{Drug name, character}
\item{n}{Number of antibiotic treatment days, 1 patient x 1 drug = 1 day}
}
}
\source{
\url{https://regionh.dk/}
}
\usage{
abx_days
}
\description{
A dataset containing antibiotic sales data from Danish hospitals 2015-2020.
}
\keyword{datasets}
