% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awr_funs.R
\name{awr_aggregate}
\alias{awr_aggregate}
\title{Aggregate antibiotic usage data by AWaRe group}
\usage{
awr_aggregate(df, atc = atc, ddd = ddd, ..., tall = TRUE)
}
\arguments{
\item{df}{Data frame}

\item{atc}{ATC code}

\item{ddd}{Amount (usually Defined Daily Doses)}

\item{...}{Grouping variables}

\item{tall}{If TRUE (default) outputs data in tall format}
}
\value{
A data frame.
}
\description{
Aggregate antibiotic usage data by AWaRe group
}
\examples{
awr_aggregate(abx_sales)
awr_aggregate(abx_sales, atc, ddd)
awr_aggregate(abx_sales, atc, addd)
awr_aggregate(abx_sales, atc, ddd, region)
awr_aggregate(abx_sales, atc, ddd, region, tall = FALSE)
awr_aggregate(abx_sales, atc, ddd, month)
awr_aggregate(abx_sales, atc, ddd, month, region)
awr_aggregate(abx_sales, atc, ddd, month, region, hospital)
}
