% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awr_funs.R
\name{awr_plot}
\alias{awr_plot}
\title{Plot AWaRe data}
\usage{
awr_plot(
  df,
  atc = atc,
  ddd = ddd,
  time = NULL,
  unit = NULL,
  ncol = NULL,
  legend.position = NULL,
  ...
)
}
\arguments{
\item{df}{Data frame.}

\item{atc}{ATC code.}

\item{ddd}{Amount.}

\item{time}{Time period.}

\item{unit}{Organisational unit.}

\item{ncol}{Integer, number of columns in faceted plots.}

\item{legend.position}{Character, where to put legend (e.g. 'none', 'right',
'bottom').}

\item{...}{Other arguments to awr_aggregate() and ggplot().}
}
\value{
A ggplot object.
}
\description{
Plot AWaRe data
}
\examples{
awr_plot(abx_sales, atc, ddd)
awr_plot(abx_sales, atc, ddd, unit = region)
awr_plot(abx_sales, atc, ddd, time = month)
awr_plot(abx_sales, atc, ddd, time = month, unit = region)
awr_plot(abx_sales, atc, ddd, time = month, unit = hospital, ncol = 4)
}
